refactor: 로컬 스토리지에 레스토랑 정보가 있으면 사용하는 로직 수정
